OboeTester: adding data path analyzer
Added BaseSineAnalyzer.
Cleaned up Activity classes.
