+++ apps/OboeTester/app/src/main/java/com/google/sample/oboe/manualtest/GlitchActivity.java
@@ -313,6 +313,10 @@ public class GlitchActivity extends AnalyzerActivity {
         closeAudio();
     }
 
+    public void stopTest() {
+        stopAudio();
+    }
+
     @Override
     boolean isOutput() {
         return false;
