+++ apps/OboeTester/app/src/main/res/layout/activity_auto_glitches.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.sample.oboe.manualtest.BaseAutoGlitchActivity">
+    tools:context="com.google.sample.oboe.manualtest.AutomatedGlitchActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
