+++ apps/OboeTester/app/src/main/res/layout/activity_data_paths.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.sample.oboe.manualtest.TestDisconnectActivity">
+    tools:context="com.google.sample.oboe.manualtest.TestDataPathsActivity">
 
     <TextView
         android:id="@+id/text_instructions"
